Move add UT into new file and revise sub UT
