+++ test/mkldnn/test_elementwise_add_mkldnn_op.py
@@ -122,23 +122,6 @@ class TestOneDNNlementwiseAddOpZeroDim3(TestOneDNNElementwiseAddOp):
         self.out = np.add(self.x, self.y)
 
 
-# Special cases for swin transformer, will ignore grad check
-class TestOneDNNlementwiseAddSrcDifferentShape(TestOneDNNElementwiseAddOp):
-    def init_input_output(self):
-        self.x = np.random.random((6, 1, 144)).astype(self.dtype)
-        self.y = np.random.random((6, 144, 1)).astype(self.dtype)
-        self.out = np.add(self.x, self.y)
-
-    def test_check_grad_normal(self):
-        pass
-
-    def test_check_grad_ingore_x(self):
-        pass
-
-    def test_check_grad_ingore_y(self):
-        pass
-
-
 ''' INT8 Tests '''
