+++ test/mkldnn/test_elementwise_sub_onednn_op.py
@@ -210,8 +210,8 @@ class TestOneDNNElementwiseSubOpZeroDim3(TestOneDNNElementwiseSubOp):
 # Special cases for swin transformer, will ignore grad check
 class TestOneDNNlementwiseSubSrcDifferentShape(TestOneDNNElementwiseSubOp):
     def init_input_output(self):
-        self.x = np.random.random((1, 4, 16, 12, 12)).astype(self.dtype)
-        self.y = np.random.random((1, 4, 1, 12, 12)).astype(self.dtype)
+        self.x = np.random.random((6, 1, 144)).astype(self.dtype)
+        self.y = np.random.random((6, 144, 1)).astype(self.dtype)
         self.out = np.subtract(self.x, self.y)
 
     def test_check_grad_normal(self):
